requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.30.0


# Bleeding edge Django
django==1.11.17 # pyup: <2

# Configuration
django-environ==0.4.4

# UI
django-crispy-forms==1.7.0
django-bootstrap-pagination==1.6.3
django_tinymce==2.6.0
django-bootstrap-breadcrumbs==0.8.2

# View
django-filter==1.1.0
django-autocomplete-light==3.2.10

# Models
django-model-utils==3.0.0
django-autoslug==1.9.3
django-teryt-tree==0.14.3
lxml==4.3.0

# Images
Pillow==4.3.0

# Password storage
argon2-cffi==16.3.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.34.0


# Python-PostgreSQL Database Adapter
psycopg2==2.7.3.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2017.3

# Redis support
django-redis==4.8.0
redis>=2.10.5

# Celery
celery==4.2.1  # pyup: < 5.0  # https://github.com/celery/celery
flower==0.9.2  # https://github.com/mher/flower

# Miscellaneous (eg. forms mixins)
django-atom==0.16.2
slugify==0.0.1


# Your custom requirements go here
django-autocomplete-light==3.2.10
django-extensions==2.0.5
mysqlclient==1.3.12
